+++ cpp/ycm/ClangCompleter/ClangHelpers.cpp
@@ -178,7 +178,7 @@ Range GetLocationExtent( CXSourceLocation source_location,
       std::string name = CXStringToString(
                            clang_getTokenSpelling( translation_unit, tokens[ i ] ) );
       Location end_location = location;
-      end_location.column_number_ += name.length();
+      end_location.column_number_ += static_cast<uint>( name.length() );
       final_range = Range( location, end_location );
       break;
     }
@@ -195,7 +195,7 @@ std::vector< CXUnsavedFile > ToCXUnsavedFiles(
   const std::vector< UnsavedFile > &unsaved_files ) {
   std::vector< CXUnsavedFile > clang_unsaved_files( unsaved_files.size() );
 
-  for ( uint i = 0; i < unsaved_files.size(); ++i ) {
+  for ( size_t i = 0; i < unsaved_files.size(); ++i ) {
     clang_unsaved_files[ i ].Filename = unsaved_files[ i ].filename_.c_str();
     clang_unsaved_files[ i ].Contents = unsaved_files[ i ].contents_.c_str();
     clang_unsaved_files[ i ].Length   = unsaved_files[ i ].length_;
@@ -213,7 +213,7 @@ std::vector< CompletionData > ToCompletionDataVector(
     return completions;
 
   completions.reserve( results->NumResults );
-  unordered_map< std::string, uint > seen_data;
+  unordered_map< std::string, size_t > seen_data;
 
   for ( uint i = 0; i < results->NumResults; ++i ) {
     CXCompletionResult completion_result = results->Results[ i ];
@@ -222,9 +222,9 @@ std::vector< CompletionData > ToCompletionDataVector(
       continue;
 
     CompletionData data( completion_result );
-    uint index = GetValueElseInsert( seen_data,
-                                     data.original_string_,
-                                     completions.size() );
+    size_t index = GetValueElseInsert( seen_data,
+                                       data.original_string_,
+                                       completions.size() );
 
     if ( index == completions.size() ) {
       completions.push_back( boost::move( data ) );
